Added self avg graphs
These graphs look at time spent with children in 97, 02, and 07 for different
types of time use and by parental risk tolerance levels.  The main thing I want
to get out of these graphs is to show that there is a big gradient, and that
the gradient is really not driven by basic care, which suggests that the
effects are not about avoiding very bad outcomes.