This reproduces my stata analysis!
This is great!  One key change that I made was to calculate my own white
standard errors using sandwich's meat and bread, but removing NAs from the
erroneous meat.  Next I will add Latex.  Also, this is a weird performance
thing because I really think it takes most of the time to just load in all the
variables and labels, and I really don't want to do that that much # Please
enter the commit message for your changes. Lines starting